requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.13
python-dotenv==1.0.0
requests==2.31.0
wolframalpha==5.0.0
wikipedia==1.4.0 